[Global Search] Fix cmd + click open in new tab for results (#183762)
## Summary

Closes https://github.com/elastic/kibana/issues/147710

This PR follows the pattern set up in
https://github.com/elastic/kibana/blob/main/x-pack/plugins/spaces/public/nav_control/components/spaces_menu.tsx#L164-L170
to allow the cmd click behavior to work for kibana global search. I'm
exploring contributing to the EUI repository in this
https://github.com/elastic/eui/pull/7788 but that avenue may not be
fruitful. This implementation is already being used in spaces and can
immediately provide value for global search with this PR merge.